New Feature:
Exporting to PDF automatically when pressing ESC or clicking CANCEL button on printer dialog.
The PDF will be written to the home folder in the subfolder "lemon-printing", and with a paper size of 72mm x 200mm.

If the dialog is accepted, then we print what the user choosed.
The user can press ENTER when dialog appearing if the desired printer is the default (or the only one).